Pull Request#1 Add unit tests

+++ src/main/java/ru/clevertec/backendtest/models/receipt/Item.java
@@ -9,7 +9,7 @@ public class Item implements Goods {
     private final Goods wrappedProduct;
     private final double cost;
 
-    public Item(Goods wrappedProduct, int amount) {
+    public Item(Goods wrappedProduct, double amount) {
         this.wrappedProduct = wrappedProduct;
         this.amount = amount;
         this.cost = setCost();
